requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altgraph==0.17.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.4
billiard==4.2.1
Brotli==1.1.0
buildozer==1.5.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels-redis==4.2.0
charset-normalizer==3.4.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.0
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.4
cssselect2==0.8.0
daphne==4.2.0
decorator==5.2.1
distlib==0.3.9
distro==1.9.0
dj-database-url==2.3.0
Django==5.2
django-braces==1.15.0
django-cloudinary-storage==0.3.0
django-crum==0.7.9
django-debug-toolbar==4.3.0
django-embed-video==1.4.9
django-extensions==4.1
django-ranged-response==0.2.0
django-redisboard==8.4.0
django-rosetta==0.10.2
django-simple-captcha==0.6.2
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.16.1
flower==2.0.1
fonttools==4.58.4
frozenlist==1.5.0
git-filter-repo==2.47.0
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
humanize==4.12.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
instaloader==4.14.1
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.8.0
kombu==5.5.4
lxml==5.3.1
magic-filter==1.0.12
matplotlib-inline==0.1.7
msgpack==1.1.1
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.3
openai==1.55.3
openpyxl==3.1.5
packaging==25.0
Panda3D==1.10.15
panda3d-gltf==1.2.1
panda3d-simplepbr==0.12.0
pandas==2.2.3
parso==0.8.4
pefile==2023.2.7
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.6
polib==1.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.2.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pydyf==0.11.0
pygame==2.6.1
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyOpenGL==3.1.7
pyOpenSSL==25.1.0
pyperclip==1.9.0
pyphen==0.17.2
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-telegram-bot==21.10
pytube==15.0.0
pytz==2025.1
PyWavefront==1.3.3
pywin32-ctypes==0.2.3
redis==5.0.4
requests==2.32.3
rfc3986==1.5.0
screeninfo==0.8.1
service-identity==24.2.0
setuptools==75.6.0
sh==2.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.0
stack-data==0.6.3
stripe==9.3.0
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
ursina==7.0.0
user-agents==2.2.0
vine==5.1.0
virtualenv==20.28.0
wcwidth==0.2.13
weasyprint==61.2
webencodings==0.5.1
yarl==1.18.3
zope.interface==7.2
zopfli==0.2.3.post1